backend-fastapi/requirements.txt
fastapi
uvicorn
psycopg2-binary
asyncpg
fastapi[all]
bcrypt
python-dotenv
sqlalchemy
alembic
python-jose

